reorganize runtime tests
