+++ example/runtimeTests/template.js
@@ -1,0 +1,15 @@
+// @flow
+
+import {BigNum} from '@emurgo/react-native-haskell-shelley'
+
+import {assert} from '../util'
+
+/**
+ * <Name>
+ */
+
+const test: () => void = async () => {
+
+}
+
+export default test
